# akshayapatra/swagger/admin-api.yaml
openapi: 3.0.3
info:
  title: Akshayapatra Admin API
  description: |
    Comprehensive Admin API for the Akshayapatra lottery platform.
    
    This API provides endpoints for administrative operations including:
    - User management with filtering, search, and pagination
    - Card management and tracking
    - Transaction monitoring and processing
    - Income and commission analytics
    - Referral system management
    - Staff role management
    - System settings configuration
    - Support and audit operations
    - Dashboard analytics and statistics
    - Payment processing and reconciliation
    
    ## Testing Mode (Current)
    **IMPORTANT**: Authentication is currently DISABLED for testing purposes.
    
    - All authentication and role checks are commented out
    - APIs use Supabase Service Role Key for full database access
    - No authentication headers required for testing
    - Mock user objects are used where user context is needed
    
    ## Authentication (When Re-enabled)
    When authentication is re-enabled, all admin endpoints will require:
    - **SuperAdmin**: Full access to all endpoints
    - **Admin**: Access to most endpoints except sensitive system operations
    - **Support**: Read access and limited write access for support operations
    - **Payments**: Access to payment and transaction related endpoints
    
    ## Required Environment Variables
    ```
    SUPABASE_SERVICE_ROLE_KEY=your_service_role_key_here
...
    ```
    
...
    
...
servers:
...
paths:
...
  /api/admin/cards/{id}:
    get:
      tags:
        - Card Management
      summary: Get card details
      description: Retrieve detailed card information including subscriptions and referral tree
      parameters:
        - name: id
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDetailsResponse'
    
    put:
      tags:
        - Card Management
      summary: Update card
      description: Update card information or suspend/unsuspend (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
      responses:
        '200':
          description: Card updated successfully
    
    delete:
      tags:
        - Card Management
      summary: Delete/deactivate card
      description: Soft delete card by deactivating (SuperAdmin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Card ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card deactivated successfully

  # Transaction Management APIs
...
components:
...
  schemas:
...
    CardDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            cardName:
              type: string
            cardholderName:
              type: string
            phoneNumber:
              type: string
            referralCode:
              type: string
            user:
              type: object
            referringCard:
              type: object
              nullable: true
            totalWalletBalance:
              type: number
            commissionWalletBalance:
              type: number
            isActive:
              type: boolean
            subscriptions:
              type: array
              items:
                type: object
            upline:
              type: array
              items:
                type: object
            downline:
              type: array
              items:
                type: object
            stats:
              type: object
            recentTransactions:
              type: array
              items:
                type: object
    UpdateCardRequest:
      type: object
      properties:
        card_name:
          type: string
        cardholder_name:
          type: string
        phone_number:
          type: string
        is_active:
          type: boolean
        suspend:
          type: boolean
          description: Whether to suspend the card
        suspension_reason:
          type: string
          description: Reason for suspension (required if suspend=true)

    # Transaction Management Schemas
...